backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
pymysql==1.1.2
alembic==1.18.1
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2